# This is a combination of 2 commits. # The first commit's message is:
moving flags to configdefaults

# The 2nd commit message will be skipped:

#	moving rest to configdefaults
