+++ theano/compile/nanguardmode.py
@@ -11,27 +11,6 @@ import theano.tensor as T
 import theano.sandbox.cuda as cuda
 from theano.compile import Mode
 
-AddConfigVar('NanGuardMode.nan_is_error',
-             "Default value for nan_is_error",
-             BoolParam(True),
-             in_c_key=False)
-
-AddConfigVar('NanGuardMode.inf_is_error',
-             "Default value for inf_is_error",
-             BoolParam(True),
-             in_c_key=False)
-
-AddConfigVar('NanGuardMode.big_is_error',
-             "Default value for big_is_error",
-             BoolParam(True),
-             in_c_key=False)
-
-AddConfigVar('NanGuardMode.action',
-             "What NanGuardMode does when it finds a problem",
-             EnumStr('raise', 'warn', 'pdb'),
-             in_c_key=False)
-
-
 logger = logging.getLogger("theano.compile.nanguardmode")
