+++ theano/gof/opt.py
@@ -1042,13 +1042,6 @@ class LocalOptimizer(object):
             (' ' * level), self.__class__.__name__, id(self)), file=stream)
 
 
-theano.configparser.AddConfigVar(
-    'metaopt.verbose',
-    "Enable verbose output for meta optimizers",
-    theano.configparser.BoolParam(False),
-    in_c_key=False)
-
-
 class LocalMetaOptimizer(LocalOptimizer):
     """
     Base class for meta-optimizers that try a set of LocalOptimizers
